# frontend/client-code-gen/fam-openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "Forest Access Management - FAM - API",
    "description" : "Forest Access Management API. Defines who has access to which applications, and the roles they operate under once access is granted.\n\nCovers both the app-access-control and admin-management surfaces that upstream FAM served from two separate APIs.",
    "contact" : {
      "name" : "Team Heartwood",
      "url" : "https://apps.nrs.gov.bc.ca/int/confluence/display/FSAST1/Team+Heartwood",
      "email" : "SIBIFSAF@Victoria1.gov.bc.ca"
    },
    "license" : {
      "name" : "Apache 2.0",
      "url" : "https://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version" : "1.0.0"
  },
  "servers" : [ {
    "url" : "/",
    "description" : "Relative to the host serving this document"
  } ],
  "security" : [ {
    "bearerAuth" : [ ]
  } ],
  "paths" : {
...
    "/fam-applications/{applicationId}/users/{userId}" : {
      "get" : {
        "tags" : [ "FAM Applications" ],
        "summary" : "Retrieve user information by user ID under an application",
        "operationId" : "get_application_user_by_id",
        "parameters" : [ {
          "name" : "applicationId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "userId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/FamUserInfoDto"
                }
              }
            }
          }
        }
      }
    },
    "/fam-applications/{applicationId}/user-role-assignment" : {
      "get" : {
        "tags" : [ "FAM Applications" ],
        "summary" : "Paged users and roles assigned within an application",
        "operationId" : "get_fam_application_user_role_assignment",
        "parameters" : [ {
          "name" : "applicationId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "pageNumber",
          "in" : "query",
          "description" : "Page number",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 1,
            "description" : "Page number",
            "minimum" : 1
          }
        }, {
          "name" : "pageSize",
          "in" : "query",
          "description" : "Number of records per page",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 50,
            "description" : "Number of records per page",
            "maximum" : 100000,
            "minimum" : 10
          }
        }, {
          "name" : "search",
          "in" : "query",
          "description" : "Search by keyword",
          "required" : false,
          "schema" : {
            "type" : "string",
            "description" : "Search by keyword",
            "maxLength" : 30,
            "minLength" : 3
          }
        }, {
          "name" : "sortOrder",
          "in" : "query",
          "description" : "Column sorting order",
          "required" : false,
          "schema" : {
            "$ref" : "#/components/schemas/SortOrder"
          }
        }, {
          "name" : "sortBy",
          "in" : "query",
          "description" : "Column to be sorted by",
          "required" : false,
          "schema" : {
            "$ref" : "#/components/schemas/UserRoleSortBy"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PagedResultsFamApplicationUserRoleAssignmentGetDto"
                }
              }
            }
          }
        }
      }
    },
    "/fam-applications/{applicationId}/user-role-assignment/export" : {
      "get" : {
        "tags" : [ "FAM Applications" ],
        "summary" : "Export user roles information by application ID",
        "operationId" : "export_application_user_roles",
        "parameters" : [ {
          "name" : "applicationId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "text/csv" : {
                "schema" : {
                  "type" : "string",
                  "format" : "byte"
                }
              }
            }
          }
        }
      }
    },
    "/external/v1/users" : {
      "get" : {
        "tags" : [ "FAM External API" ],
        "summary" : "Search FAM users",
        "description" : "Search FAM users associated with the calling application.",
        "operationId" : "user_search",
        "parameters" : [ {
          "name" : "page",
          "in" : "query",
          "description" : "Page number - 1 index",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "description" : "Page number - 1 index",
            "minimum" : 1
          }
        }, {
          "name" : "size",
          "in" : "query",
          "description" : "Number of records per page",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "description" : "Number of records per page",
            "maximum" : 100,
            "minimum" : 10
          }
        }, {
          "name" : "idpType",
          "in" : "query",
          "description" : "Identity provider type. Available values: IDIR, BCEID, BCSC",
          "required" : false,
          "schema" : {
            "$ref" : "#/components/schemas/IdpType"
          }
        }, {
          "name" : "idpUsername",
          "in" : "query",
          "description" : "Username from the identity provider",
          "required" : false,
          "schema" : {
            "type" : "string",
            "description" : "Username from the identity provider",
            "maxLength" : 20,
            "minLength" : 0
          }
        }, {
          "name" : "firstName",
          "in" : "query",
          "description" : "User's first name",
          "required" : false,
          "schema" : {
            "type" : "string",
            "description" : "User's first name",
            "maxLength" : 50,
            "minLength" : 0
          }
        }, {
          "name" : "lastName",
          "in" : "query",
          "description" : "User's last name",
          "required" : false,
          "schema" : {
            "type" : "string",
            "description" : "User's last name",
            "maxLength" : 50,
            "minLength" : 0
          }
        }, {
          "name" : "role",
          "in" : "query",
          "description" : "Role codes to filter by, e.g. ILCR_SUBMITTER",
          "required" : false,
          "schema" : {
            "type" : "array",
            "description" : "Role codes to filter by, e.g. ILCR_SUBMITTER",
            "items" : {
              "type" : "string",
              "description" : "Role codes to filter by, e.g. ILCR_SUBMITTER",
              "maxLength" : 25,
              "minLength" : 0
            },
            "maxItems" : 5,
            "minItems" : 0
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ExtUserSearchPagedResults"
                }
              }
            }
          }
        }
      }
    },
...
    "/external/v1/users/identity/idir/search" : {
      "get" : {
        "tags" : [ "FAM External API" ],
        "summary" : "Search IDIR users",
        "description" : "Search IDIR user identities through IDIM.",
        "operationId" : "search_idim_idir_users",
        "parameters" : [ {
          "name" : "firstName",
          "in" : "query",
          "description" : "IDIR first name search value (min 2 chars)",
          "required" : false,
          "schema" : {
            "type" : "string",
            "description" : "IDIR first name search value (min 2 chars)",
            "maxLength" : 50,
            "minLength" : 0
          }
        }, {
          "name" : "lastName",
          "in" : "query",
          "description" : "IDIR last name search value (min 2 chars)",
          "required" : false,
          "schema" : {
            "type" : "string",
            "description" : "IDIR last name search value (min 2 chars)",
            "maxLength" : 50,
            "minLength" : 0
          }
        }, {
          "name" : "userId",
          "in" : "query",
          "description" : "IDIR user id search value (min 2 chars)",
          "required" : false,
          "schema" : {
            "type" : "string",
            "description" : "IDIR user id search value (min 2 chars)",
            "maxLength" : 20,
            "minLength" : 0
          }
        }, {
          "name" : "pageSize",
          "in" : "query",
          "description" : "Number of records to return",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "description" : "Number of records to return",
            "maximum" : 500,
            "minimum" : 10
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/IdimProxyIdirUsersSearchResultDto"
                }
              }
            }
          }
        },
        "deprecated" : true
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "UserType" : {
        "type" : "string",
        "enum" : [ "I", "B", "CD", "CT", "CP" ]
      },
...
      "FamApplicationDto" : {
        "properties" : {
          "application_id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "application_name" : {
            "type" : "string"
          },
          "application_description" : {
            "type" : "string"
          }
        },
        "required" : [ "application_description", "application_id", "application_name" ]
      },
      "FamApplicationUserRoleAssignmentGetDto" : {
        "properties" : {
          "user_role_xref_id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "user_id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "role_id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "user" : {
            "$ref" : "#/components/schemas/FamUserInfoDto"
          },
          "role" : {
            "$ref" : "#/components/schemas/FamRoleWithClientDto"
          },
          "create_date" : {
            "type" : "string",
            "format" : "date-time"
          },
          "expiry_date" : {
            "type" : "string",
            "format" : "date-time"
          }
        },
        "required" : [ "create_date", "role", "role_id", "user", "user_id", "user_role_xref_id" ]
      },
      "FamForestClientDto" : {
        "properties" : {
          "client_name" : {
            "type" : "string"
          },
          "forest_client_number" : {
            "type" : "string"
          },
          "status" : {
            "$ref" : "#/components/schemas/FamForestClientStatusDto"
          }
        },
        "required" : [ "forest_client_number" ]
      },
      "FamForestClientStatusDto" : {
        "properties" : {
          "status_code" : {
            "$ref" : "#/components/schemas/FamForestClientStatusType"
          },
          "description" : {
            "type" : "string"
          }
        }
      },
      "FamForestClientStatusType" : {
        "type" : "string",
        "enum" : [ "A", "I" ]
      },
      "FamRoleMinDto" : {
        "properties" : {
          "role_name" : {
            "type" : "string"
          },
          "role_type_code" : {
            "type" : "string"
          },
          "application" : {
            "$ref" : "#/components/schemas/FamApplicationDto"
          }
        },
        "required" : [ "application", "role_name", "role_type_code" ]
      },
      "FamRoleWithClientDto" : {
        "properties" : {
          "role_id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "role_name" : {
            "type" : "string"
          },
          "role_type_code" : {
            "type" : "string"
          },
          "display_name" : {
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          },
          "application" : {
            "$ref" : "#/components/schemas/FamApplicationDto"
          },
          "forest_client" : {
            "$ref" : "#/components/schemas/FamForestClientDto"
          },
          "parent_role" : {
            "$ref" : "#/components/schemas/FamRoleMinDto"
          }
        },
        "required" : [ "application", "description", "forest_client", "role_id", "role_name", "role_type_code" ]
      },
      "FamUserInfoDto" : {
        "properties" : {
          "user_name" : {
            "type" : "string"
          },
          "user_type" : {
            "$ref" : "#/components/schemas/FamUserTypeDto"
          },
          "first_name" : {
            "type" : "string"
          },
          "last_name" : {
            "type" : "string"
          },
          "email" : {
            "type" : "string"
          }
        },
        "required" : [ "user_name", "user_type" ]
      },
...
      "FamUserTypeDto" : {
        "properties" : {
          "code" : {
            "$ref" : "#/components/schemas/UserType"
          },
          "description" : {
            "type" : "string"
          }
        },
        "required" : [ "code", "description" ]
      },
...
      "IdimProxyIdirUserSearchItemDto" : {
        "properties" : {
          "userId" : {
            "type" : "string"
          },
          "guid" : {
            "type" : "string"
          },
          "firstName" : {
            "type" : "string"
          },
          "lastName" : {
            "type" : "string"
          },
          "email" : {
            "type" : "string"
          }
        },
        "required" : [ "email", "firstName", "guid", "lastName", "userId" ]
      },
      "IdimProxyIdirUsersSearchResultDto" : {
        "properties" : {
          "totalItems" : {
            "type" : "integer",
            "format" : "int32"
          },
          "pageSize" : {
            "type" : "integer",
            "format" : "int32"
          },
          "items" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/IdimProxyIdirUserSearchItemDto"
            }
          }
        },
        "required" : [ "items", "pageSize", "totalItems" ]
      },
...
      "SortOrder" : {
        "type" : "string",
        "description" : "Column sorting order",
        "enum" : [ "asc", "desc" ]
      },
      "UserRoleSortBy" : {
        "type" : "string",
        "description" : "Column to be sorted by",
        "enum" : [ "create_date", "user_name", "user_type_code", "email", "full_name", "role_display_name", "forest_client_number" ]
      },
      "PageResultMeta" : {
        "properties" : {
          "total" : {
            "type" : "integer",
            "format" : "int64",
            "description" : "Total records counts for query conditions"
          },
          "number_of_pages" : {
            "type" : "integer",
            "format" : "int32",
            "description" : "Total pages for query records"
          },
          "page_number" : {
            "type" : "integer",
            "format" : "int32",
            "description" : "Page number"
          },
          "page_size" : {
            "type" : "integer",
            "format" : "int32",
            "description" : "Number of records per page"
          }
        },
        "required" : [ "number_of_pages", "page_number", "page_size", "total" ]
      },
      "PagedResultsFamApplicationUserRoleAssignmentGetDto" : {
        "properties" : {
          "meta" : {
            "$ref" : "#/components/schemas/PageResultMeta"
          },
          "results" : {
            "type" : "array",
            "description" : "Paged results",
            "items" : {
              "$ref" : "#/components/schemas/FamApplicationUserRoleAssignmentGetDto"
            }
          }
        },
        "required" : [ "meta", "results" ]
      },
      "IdpType" : {
        "type" : "string",
        "description" : "Identity provider type. Available values: IDIR, BCEID, BCSC",
        "enum" : [ "IDIR", "BCEID", "BCSC" ]
      },
      "ExtApplicationUserSearchGetDto" : {
        "properties" : {
          "firstName" : {
            "type" : "string"
          },
          "lastName" : {
            "type" : "string"
          },
          "idpUsername" : {
            "type" : "string"
          },
          "idpUserGuid" : {
            "type" : "string"
          },
          "idpType" : {
            "$ref" : "#/components/schemas/IdpType"
          },
          "roles" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/ExtRoleWithScopeDto"
            }
          }
        }
      },
      "ExtPageResultMeta" : {
        "properties" : {
          "total" : {
            "type" : "integer",
            "format" : "int64"
          },
          "pageCount" : {
            "type" : "integer",
            "format" : "int32"
          },
          "page" : {
            "type" : "integer",
            "format" : "int32"
          },
          "size" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
      "ExtRoleWithScopeDto" : {
        "properties" : {
          "applicationName" : {
            "type" : "string"
          },
          "roleName" : {
            "type" : "string"
          },
          "roleDisplayName" : {
            "type" : "string"
          },
          "scopeType" : {
            "$ref" : "#/components/schemas/ScopeType"
          },
          "value" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          }
        }
      },
      "ExtUserSearchPagedResults" : {
        "properties" : {
          "meta" : {
            "$ref" : "#/components/schemas/ExtPageResultMeta"
          },
          "users" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/ExtApplicationUserSearchGetDto"
            }
          }
        }
      },
      "ScopeType" : {
        "type" : "string",
        "enum" : [ "FOREST_CLIENT" ]
      },
...
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "type" : "http",
        "description" : "Access token issued by BC Gov SSO (Keycloak).",
        "scheme" : "bearer",
        "bearerFormat" : "JWT"
      },
      "apiKey" : {
        "type" : "apiKey",
        "description" : "Shared secret for the scheduled user-information refresh.",
        "name" : "X-API-Key",
        "in" : "header"
      }
    }
  }
}